opt: Split ExprView methods into new file
Move ExprView formatting-related methods into new expr_view_format.go
file.

Release note: None
